commented and restructured non-api code
