+++ main/templates/main/query.html
@@ -63,7 +63,7 @@
 									<td> {{ player.stats.mpg }} </td>
 									<td> {{ player.stats.ft }} </td>
 									<td> {{ player.stats.gp }} </td>
-									<td> {{ player.stats.to }} </td>
+									<td> {{ player.stats.topg }} </td>
 								</tr>
 							</tbody>
 						</table>
