+++ main/views/api.py
@@ -1,0 +1,68 @@
+from rest_framework.views import APIView
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.authtoken.models import Token
+from rest_framework.response import Response
+from rest_framework.exceptions import NotFound, ParseError
+
+from main.views.helpers.current_season_stats_job import CurrentSeasonStatsJob
+
+
+class PlayersList(APIView):
+    permission_classes = (IsAuthenticated,)
+
+    def get(self, request, *args, **kwargs):
+        query_set = self.get_players()
+        serializer = PlayerSerializer(query_set, many=True)
+        resp_data = {'count' : len(query_set), 'players' : serializer.data}
+        return Response(resp_data)
+
+    def get_players(self):
+        query = Player.objects.all()
+        name_ = self.request.query_params.get('name')
+        draft_year_ = self.request.query_params.get('draft_year')
+        position_ = self.request.query_params.get('position')
+        if name_:
+            query = query.filter(name__iexact = name_)
+        if draft_year_:
+            if not draft_year_.isdigit():
+                raise ParseError('Draft year must be a positive integer')
+            query = query.filter(year_enter_league = draft_year_)
+        if position_:
+            if position_.lower() not in ['center', 'forward', 'guard']:
+                raise NotFound('Incorrect position')
+            query = query.filter(position__iexact = position_)
+        return query
+
+class PastStatisticsList(APIView):
+    permission_classes = (IsAuthenticated,)
+
+    def get(self, request, *args, **kwargs):
+        query_set = self.get_stats()
+        serializer = StatisticsSerializer(query_set, many=True)
+        resp_data = {'count' : len(query_set), 'stats' : serializer.data}
+        return Response(resp_data)
+
+    def get_stats(self):
+        name_ = self.request.query_params.get('name')
+        if not name_:
+            raise ParseError('Missing \'name\'')
+        season_ = self.request.query_params.get('season')
+        query = Statistics.objects.filter(name__iexact = name_)
+        if season_:
+            season_ = str(season_)
+            if not season_.isdigit():
+                raise ParseError('Season must be a positive integer')
+            formatted_season = '%s-%s' % (season_, str(int(season_) + 1)[-2:])
+            query = query.filter(season = formatted_season)
+        return query
+
+class CurrentStatistics(PastStatisticsList):
+    def get_stats(self):
+        current_season_stats = CurrentSeasonStatsJob().get()
+        name_ = self.request.query_params.get('name')
+        if name_:
+            name_ = name_.lower()
+            stats = filter(lambda s: s.name.lower() == name_, current_season_stats)
+            return stats
+        else:
+            return current_season_stats
